Moved CSV generation to sf-utils.ts file
