+++ packages/destination-actions/src/destinations/salesforce/sf-object-to-shape.ts
@@ -95,13 +95,3 @@ export const mapObjectToShape = (payload: GenericPayload, sobject: string): Gene
 
   return shapeFunction(payload)
 }
-
-// Our key names are in snake case, but Salesforce's field names are in camel case.
-// I.E. 'last_name' is our key, but 'LastName' is the Salesforce field name.
-export const snakeCaseToCamelCase = (key: string): string => {
-  const tokens = key.split('_')
-  tokens.forEach((token, i, arr) => {
-    arr[i] = token.charAt(0).toUpperCase() + token.slice(1)
-  })
-  return tokens.join('')
-}
